refactor: split gameService into multiple services
